experimental addition in macro flag USE_GLLISTS:
horizontal contours are stored in glLists rather than as lists of vectors.
This turns over memory management for these objects to the ogl library, and
frees vis5d memory for other uses.  Compiling with this flag will break
save and restore functions for hslice.  This flag should be considered
experimental at this time.

Jim