gtk/graph_labels.c: silence newer gcc compilers by adding break statements
Newer gcc compilers are more strict when handling labels and consider
it an error to have a label at the end of a block. For example:

  switch (var) {
    case '1':
       /* do something */
    break;
    default:
  }

will cause newer gcc to throw an error because of the dangling 'default:'
label.

Add a break statement after the default label to fix it.