configure.in: set "tarname" field in AC_INIT macro
Since this field was not set, the PACKAGE variable which is set by
AM_INIT_AUTOMAKE was set to a cleansed version of the "package" field
in the AC_INIT macro. i.e. when the package name 'vis5d+' was passed
as the first argument to AC_INIT, PACKAGE was set to 'vis5d-'.

This is important because files are installed to $prefix/include/$PACKAGE
and $prefix/share/$PACKAGE, and so these installation directories were
being improperly named.